move inputs to their own folder
