+++ app/links/components/LinkForm.tsx
@@ -1,11 +1,11 @@
 import Heading from "@/app/ui/components/Heading"
 import Text from "@/app/ui/components/Text"
-import { AllMenuList } from "@/app/ui/components/AllMenuList"
+import { AllMenuList } from "@/app/ui/inputs/AllMenuList"
 import { Edge } from "@atlaskit/pragmatic-drag-and-drop-hitbox/closest-edge"
 import { DropIndicator } from "@atlaskit/pragmatic-drag-and-drop-react-drop-indicator/box"
 import { DragHandleButton } from "@atlaskit/pragmatic-drag-and-drop-react-accessibility/drag-handle-button"
 import { RefObject } from "react"
-import { InputField } from "@/app/ui/components/InputField"
+import { InputField } from "@/app/ui/inputs/InputField"
 import { useForm } from "react-hook-form"
 import { zodResolver } from "@hookform/resolvers/zod"
 import { formSchema } from "@/app/ui/demos/FormDemo"
