+++ app/profile-details/page.tsx
@@ -1,6 +1,6 @@
 import Heading from "@/app/ui/components/Heading"
 import Text from "@/app/ui/components/Text"
-import { Button } from "@/app/ui/components/Button"
+import { Button } from "@/app/ui/inputs/Button"
 import Image from "next/image"
 
 export default function Page() {
